Extract labeling code to its own module
